% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata_helpers.R
\name{generate_garbage_values}
\alias{generate_garbage_values}
\title{Generate garbage values}
\usage{
generate_garbage_values(garbage_type, garbage_row, variable_type, n)
}
\arguments{
\item{garbage_type}{Character. "garbage_low", "garbage_high", "garbage_future", etc.}

\item{garbage_row}{Data frame row. Contains recStart with interval notation for garbage range.}

\item{variable_type}{Character. "continuous", "date", etc.}

\item{n}{Integer. Number of values to generate.}
}
\value{
Vector of garbage values.
}
\description{
Generates implausible values for garbage based on type.
Helper function for make_garbage().
}
\seealso{
Other mockdata-helpers: 
\code{\link{apply_garbage}()},
\code{\link{apply_missing_codes}()},
\code{\link{apply_rtype_defaults}()},
\code{\link{extract_distribution_params}()},
\code{\link{extract_proportions}()},
\code{\link{get_cycle_variables}()},
\code{\link{get_raw_variables}()},
\code{\link{get_variable_details}()},
\code{\link{has_garbage}()},
\code{\link{make_garbage}()},
\code{\link{sample_with_proportions}()}
}
\concept{mockdata-helpers}
