% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_spec.R
\name{is_mock_spec}
\alias{is_mock_spec}
\title{Check whether an object is a MockData specification}
\usage{
is_mock_spec(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical scalar.
}
\description{
Check whether an object is a MockData specification
}
\examples{
spec <- mock_spec()
is_mock_spec(spec)
is_mock_spec(list())

}
