% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MockData-package.R
\docType{package}
\name{MockData-package}
\alias{MockData}
\alias{MockData-package}
\title{MockData: Generate Mock Data from Metadata Specifications}
\description{
Generates mock testing data from small metadata specifications, supplied as in-memory data frames or CSV files. Supports categorical, continuous, date, garbage-data, and survival-style variables for testing data pipelines and documentation examples. Designed for recodeflow-style metadata while also supporting lightweight standalone use.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://big-life-lab.github.io/MockData/}
  \item \url{https://github.com/Big-Life-Lab/MockData}
  \item Report bugs at \url{https://github.com/Big-Life-Lab/MockData/issues}
}

}
\author{
\strong{Maintainer}: Douglas Manuel \email{dmanuel@ohri.ca}

Authors:
\itemize{
  \item Juan Li \email{juli@ohri.ca}
}

Other contributors:
\itemize{
  \item recodeflow contributors [contributor]
}

}
\keyword{internal}
