% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_mockdata_extensions.R
\name{print.mockdata_validation_result}
\alias{print.mockdata_validation_result}
\title{Print MockData Validation Results}
\usage{
\method{print}{mockdata_validation_result}(x, ...)
}
\arguments{
\item{x}{mockdata_validation_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Print MockData Validation Results
}
