% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_mockdata_extensions.R
\name{validate_foreign_keys}
\alias{validate_foreign_keys}
\title{Validate Foreign Key Relationships}
\usage{
validate_foreign_keys(variables, variable_details)
}
\description{
Validate Foreign Key Relationships
}
\keyword{internal}
