% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_mockdata_extensions.R
\name{validate_garbage}
\alias{validate_garbage}
\title{Validate Garbage Data Parameters}
\usage{
validate_garbage(variables)
}
\description{
Validate Garbage Data Parameters
}
\keyword{internal}
