% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_mockdata_extensions.R
\name{validate_proportions}
\alias{validate_proportions}
\title{Validate Proportion Sums}
\usage{
validate_proportions(variable_details, mode)
}
\description{
Validate Proportion Sums
}
\keyword{internal}
