% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_mockdata_extensions.R
\name{validate_versioning}
\alias{validate_versioning}
\title{Validate MockData Versioning Fields}
\usage{
validate_versioning(variables)
}
\description{
Validate MockData Versioning Fields
}
\keyword{internal}
